Make the skill queue countdown timer update, and keep in synch with the current skill training countdown.  Rip the guts out of MTCountdown class.

+++ src/Core/Character/Character.m
@@ -166,8 +166,10 @@
 						 secondary:(NSInteger)secondary 
 					   skillPoints:(NSInteger)skillPoints
 {
-	CGFloat spPerSecond = ((((CGFloat)attributeTotals[primary]) + (((CGFloat)attributeTotals[secondary]) / 2.0)) / 60.0);
+	CGFloat spPerSecond = (((attributeTotals[primary]) + ((attributeTotals[secondary]) / 2.0)) / 60.0);
+	
 	CGFloat trainingTime = ((CGFloat)skillPoints / spPerSecond);
+	
 	return xlround(trainingTime);
 }
